Tests/LibWeb/Text/expected/html-parser-script-created-parser-many-flat-siblings.txt

PASS
parser appended every sibling using the append sentinel
